F1313 freeze

+++ src/src/vdp/gwenesis_vdp.h
@@ -18,6 +18,7 @@ __license__ = "GPLv3"
 */
 #ifndef _gwenesis_vdp_H_
 #define _gwenesis_vdp_H_
+#include <stdint.h>
 
 #pragma once
 
@@ -129,7 +130,7 @@ void gwenesis_vdp_write_memory_8(unsigned int address, unsigned int value);
 void gwenesis_vdp_write_memory_16(unsigned int address, unsigned int value);
 
 void gwenesis_vdp_set_buffers(unsigned char *screen_buffer, unsigned char *scaled_buffer);
-void gwenesis_vdp_set_buffer(unsigned short *ptr_screen_buffer);
+void gwenesis_vdp_set_buffer(uint8_t *ptr_screen_buffer);
 void gwenesis_vdp_render_line(int line);
 
 void gwenesis_vdp_render_config();
