+++ src/src/vdp/gwenesis_vdp_gfx.c
@@ -126,7 +126,7 @@ void gwenesis_vdp_set_buffers(unsigned char *screen_buffer, unsigned char *scale
     scaled_screen = scaled_buffer;
 }
 //embedded
-void gwenesis_vdp_set_buffer(unsigned short *ptr_screen_buffer)
+void gwenesis_vdp_set_buffer(uint8_t *ptr_screen_buffer)
 {
     screen_buffer_line = ptr_screen_buffer;
     screen_buffer = ptr_screen_buffer;
